requirements.txt
# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
pydantic==2.10.4
pydantic-settings==2.7.1

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
geoalchemy2==0.15.2

# Auth
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
argon2-cffi==23.1.0
bcrypt==4.2.1
httpx==0.28.1

# Utils
email-validator==2.2.0

# Testing
pytest==8.3.4
pytest-asyncio==0.25.2
pytest-cov==6.0.0
aiosqlite==0.20.0

# Dev
ruff==0.8.6
